refactor(state): extract registry normalization
