+++ src/lib/state/registry-normalization.ts
@@ -1,0 +1,31 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import { isObjectRecord } from "../core/json-types";
+import type { SandboxEntry, SandboxRegistry } from "./registry";
+
+export function sandboxRegistryEntries(data: SandboxRegistry): Array<[string, SandboxEntry]> {
+  const sandboxes = isObjectRecord(data.sandboxes) ? data.sandboxes : {};
+  return Object.entries(sandboxes).filter((entry): entry is [string, SandboxEntry] =>
+    isSandboxEntryLike(entry[0], entry[1]),
+  );
+}
+
+function isSandboxEntryLike(name: string, entry: unknown): entry is SandboxEntry {
+  return (
+    isObjectRecord(entry) &&
+    typeof entry.name === "string" &&
+    entry.name === name &&
+    entry.name.trim().length > 0
+  );
+}
+
+export function retainedDefaultSandbox(
+  defaultSandbox: unknown,
+  sandboxes: Record<string, SandboxEntry>,
+): string | null {
+  if (typeof defaultSandbox !== "string") return null;
+  const entry = sandboxes[defaultSandbox];
+  if (!entry || entry.pendingRouteReservation === true) return null;
+  return defaultSandbox;
+}
